requirements.txt
streamlit
pandas
numpy
plotly
requests
scikit-learn
yfinance
matplotlib
seaborn
polygon-api-client